// src/app/Models/MenuItems.ts
export class MenuItem{
    id: string = '';
    itemName: string = '';
    price: number = 0;
    image: string = '';
    description: string = '';
    quantity: number = 0;
    categoryId: number = 0;
}

export class postMenuItem{
    itemName: string = '';
    price: number = 0;
    image: string = '';
    description: string = '';
    quantity: number = 0;
    categoryId: number = 0;
}